requirements.txt
altgraph==0.17
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.0.8
django-ckeditor==5.9.0
django-js-asset==1.2.2
future==0.18.2
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pefile==2019.4.18
Pillow==7.2.0
psycopg2==2.8.5
PyInstaller==3.6
pylint==2.5.3
pytz==2020.1
pywin32-ctypes==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
